+++ src/jsonvar/JsonVar.cpp
@@ -23,7 +23,7 @@ namespace zetjsoncpp {
 
 		case JSON_VAR_TYPE_VECTOR_OF_BOOLEANS: return "ArrayBooleanJsonVar";
 			//case ARRAY_INT32_TYPE: return "ARRAY INT32 TYPE";
-		case JSON_VAR_TYPE_VECTOR_OF_NUMBERS: return "ArrayJsonVarNumber";
+		case JSON_VAR_TYPE_VECTOR_OF_NUMBERS: return "ArrayNumberJsonVar";
 		case JSON_VAR_TYPE_VECTOR_OF_STRINGS: return "ArrayStringJsonVar";
 		case JSON_VAR_TYPE_VECTOR_OF_OBJECTS: return "ArrayObjectJsonVar";